+++ cli/src/conf/mod.rs
@@ -39,17 +39,11 @@ pub(crate) enum ConfCommand {
         )]
         sui_wallet_path: Option<PathBuf>,
         #[arg(
-            long = "sui.basic-auth-user",
-            help = "Set an user for basic authentication to the Sui node",
-            value_name = "USER"
+            long = "sui.rpc-url",
+            help = "Set a custom RPC URL for the Sui node",
+            value_name = "URL"
         )]
-        sui_auth_user: Option<String>,
-        #[arg(
-            long = "sui.basic-auth-password",
-            help = "Set a password for basic authentication to the Sui node",
-            value_name = "PASSWORD"
-        )]
-        sui_auth_password: Option<String>,
+        sui_rpc_url: Option<reqwest::Url>,
         #[arg(
             long = "nexus.objects",
             help = "Path to a TOML file containing Nexus objects",
@@ -98,8 +92,7 @@ pub(crate) async fn handle(command: ConfCommand) -> AnyResult<(), NexusCliError>
         ConfCommand::Set {
             sui_net,
             sui_wallet_path,
-            sui_auth_user,
-            sui_auth_password,
+            sui_rpc_url,
             nexus_objects_path,
             generate_identity_key,
             conf_path,
@@ -107,8 +100,7 @@ pub(crate) async fn handle(command: ConfCommand) -> AnyResult<(), NexusCliError>
             set_nexus_conf(
                 sui_net,
                 sui_wallet_path,
-                sui_auth_user,
-                sui_auth_password,
+                sui_rpc_url,
                 nexus_objects_path,
                 generate_identity_key,
                 conf_path,